Devise Gem Added, Installed and Setup for Users
Devise gem installed and added to the Gemfile.

Updated the generated initializer config email address and uncommented
it to conan@example.com

Database migratiions performed so that Devise is integrated into the application
schema along with the new User model.

Uncommented the navigational formats line to prevent potential isuess.

Users can be created via the sign up page and appear in the database and
can be viewed in the Rails console.