Enhancing Devise Authentication in the Views
Added a Navbar to the application view page for authenication purposes.

Navbar contains links to signing up, logging in, editing the profile and
logging out.

Notice and alerts set up also in the application view page.

Link to ideas will be updated to Nebula Thoughts.

Added an authentication requirement to the application controller before
actions are performed by a user.