Phlex Added and First Phlex Card Component Implemnented
Added the phlex-rails gem to the Gemfile, installed it and ran the install
command to ensure all necessary files are created and setup for Phlex to run efficiently.

Generated Phlex index view template file along with an application view file
for being a Phlex class for all Phlex views.

Generated a components folder and in particular a card component for what
 will eventually form Phlex view cards throughout the application, for now
 it is just rendered in the homepage.