Added Cloudinary Video to the Homepage
Cloudinary video of a turtle swimming added to the homepage. Transformation
styling and settings added to make the video bespoke.

The video plays once clicked for a 10 second duration before looping and
playing again repeatedly.

Border styling enacted to differentiate the video from other homepage content.
In the near future I will look to add general view page styling for separating
content.

Fallback content in place also in the cases of an incompatible browser and/or
the Cloudinary video being unavailable.