Homepage Updated, Flowbite Added, NPM and Node Updated
Application homepage is the primary reason for this branch and commit in
particular, the homepage required appearance as it was appearing disorganised.

Added Flowbite, a Tailwind component builder for helping with building simple
Tailwind for now. Also added in a Flowbite navbar in a new shared navbar partial
view file.

Updated NPM and Node to the latest versions (10.5.2 and 21.7.3 respectively).

Refactored application layout view file to remove redundant code and incorrect
indentations, neater layout now within in a local server.