Nebula Thoughts Implemented, Tailwind Flowbite Cards also
Nebula Thoughts added to the navbar, Nebula Thoughts simultaneously added which is
a page with room for experimentation with the new cards(just as the application is).

Tailwind Flowbite cards added to the application for the Nebula thoughts page.
Relevant terminology text in the cards added to the application and lorem ipsum with it.

CSS added into the base CSS application file for the newly introduced Nebula cards.

Nebula controller added to with show and thoughts actions.