Nebula Model Expanded Including New Method, Galaxy Model Created
Intention is to gradually expand the scope of what Nebula means to the
application, initially in this commit it is to start the growth of Nebula
with attributes and associations, with a distance method also included
and comments to clarify its existence.

Galaxy model also created for an initial belongs_to nebula association from
nebula to galaxy in the nebula model.