ActiveAdmin Setup and Able to Create Admin Admin Users
ActiveAdmin gem and installation performed, with also using the sassc-rails
gem to ensure the full admin interface is acesssible and not experiencing
errors.

Configured the admin dashboard to display a welcome message upon visiting
the dashboard.

Added a Ransack method to the new admin model to resolve a Ransack Rails error.

Have tried and succesfully managed to create a new admin user that is also
displayed in the admin user records.