Turbo Introduced and Supernova Setup, Several Minor Updates also made
Turbo added to the Nebula MVC, not working efficiently as of this commit.

Added the Supernova MVC and like Nebula introduced the plural of Supernovae
via the inflections.rb file. Turbo added to the Supernova system, and is working
in the new Supernova form. Many actions setup in the new Supernovae controller, destroy
to be implemented soon to complete the CRUD actions for Supernovae.

Added the Nebula about page and made it link via a Rails link_to for the page.

Routes updated to include the new Nebula about view and Supernovae resources along
with for the link from the homepage to the Supernova index page.

Added seeds for Supernovae and Nebulae.