Astro Images MVC Setup, Can be Added
Looking to add user input/their ability to add to the application.
Adding astrology images seems to be a suitable and relevant to the project
approach to this.

A astro_image model was initially created with an image required, form view
built around this to require a name, description and image attached from the
user's local system to submit to the astrology images records.

Added a link to the Navbar partial so users can access initially the Nebula index,
then click the link to access the new astrology image form page.

Limited in the routes for resources for the astro_images to only contain the actions
of index, show, new and create as they are the only ones needed at this point, as this
is not an overly complex MVC architecture for the application. The controller actions
only contain new, create and index along with a private method for deciding the controller
parameters.