Added Astro Images Show Page and Links to the First Blob
Added show action to the astro_images controller.

Added a new astro_images show page and link to via the entirety of the blob
record for each index astro_images image.

Show page features same information and image as that of an image in the astro_images
index view page.