simple_calendar Gem Setup and Implementation into the Homepage
Latest version of the simple_calendar gem added to the application,
added it to the homepage view.

Created a meeting model and pages controller to handle the calendar implementation.

The homepage now shows a calendar where likely in the next PR users will be able
to select a day from the calendar and create a meeting.

Commented out meeting path link in the month_calendar partial until this is
initialised and the relevant view created in the future.